webeng2/6. practice

+++ 2020-21-2/webeng2/mytracks/app/Http/Controllers/ProjectController.php
@@ -5,19 +5,21 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ProjectFormRequest;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
     public function index()
     {
-        $projects = Project::all();
+        // $projects = Project::all();
         return view('projects.list', [
-            'projects'  => $projects,
+            'projects'  => Auth::user()->projects,
         ]);
     }
 
     public function show(Project $project)
     {
+        $this->authorize('access', $project);
         // $project = Project::find($id);
         return view('projects.show', [
             'project' => $project,
@@ -32,6 +34,7 @@ class ProjectController extends Controller
     public function store(ProjectFormRequest $request)
     {
         $validated_data = $request->validated();
+        $validated_data['user_id'] = Auth::id();
         Project::create($validated_data);
         // $validated_data // Database save
         // return $this->index();
@@ -40,6 +43,7 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
+        $this->authorize('access', $project);
         // $project = Project::find($id);
         return view('projects.edit', [
             'project' => $project,
@@ -48,6 +52,7 @@ class ProjectController extends Controller
 
     public function update(Project $project, ProjectFormRequest $request)
     {
+        $this->authorize('access', $project);
         $validated_data = $request->validated();
         // $project = Project::find($id);
         $project->update($validated_data);
@@ -57,6 +62,7 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('access', $project);
         // Project::destroy($id);
         $project->delete();
         return redirect()->route('projects.index');
